refactor: Use a limited amount of tasks
