+++ src/main.rs
@@ -9,11 +9,14 @@ mod graceful_shutdown;
 mod logger;
 
 mod intermediate_rep;
+mod task_processing;
 mod util;
 
+use base64::{engine::general_purpose::STANDARD as BASE64, Engine as _};
 use beam_lib::{TaskRequest, TaskResult, MsgId};
 use laplace_rs::ObfCache;
 use tokio::sync::Mutex;
+use task_processing::TaskQueue;
 
 use crate::util::{is_cql_tampered_with, obfuscate_counts_mr};
 use crate::{config::CONFIG, errors::FocusError};
@@ -27,11 +30,7 @@ use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 use std::{process::exit, time::Duration};
 
-use base64::{engine::general_purpose, Engine as _};
-
 use serde::{Deserialize, Serialize};
-use serde_json::from_slice;
-
 use tracing::{debug, error, warn};
 
 // result cache
@@ -40,7 +39,6 @@ type Report = String;
 type Created = std::time::SystemTime; //epoch
 type BeamTask = TaskRequest<String>;
 type BeamResult = TaskResult<beam_lib::RawString>;
-type Shared<T> = Arc<Mutex<T>>;
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 struct Metadata {
@@ -77,13 +75,8 @@ pub async fn main() -> ExitCode {
 }
 
 async fn main_loop() -> ExitCode {
-    let obf_cache = Arc::new(Mutex::new(ObfCache {
-        cache: HashMap::new(),
-    }));
-
-    let report_cache: Shared<ReportCache> = Default::default();
-    let mut seen_tasks = Default::default();
-
+    // TODO: The report cache init should be an fn on the cache
+    let mut report_cache: ReportCache = Default::default();
     if let Some(filename) = CONFIG.queries_to_cache_file_path.clone() {
         let lines = util::read_lines(filename.clone().to_string());
         match lines {
@@ -93,7 +86,7 @@ async fn main_loop() -> ExitCode {
                         warn!("A line in the file {} is not readable", filename);
                         continue;
                     };
-                    report_cache.lock().await.cache.insert(ok_line, ("".into(), UNIX_EPOCH));
+                    report_cache.cache.insert(ok_line, ("".into(), UNIX_EPOCH));
                 }
             }
             Err(_) => {
@@ -103,6 +96,8 @@ async fn main_loop() -> ExitCode {
         }
     }
 
+    let mut seen_tasks = Default::default();
+    let mut task_queue = task_processing::spawn_task_workers(report_cache);
     let mut failures = 0;
     while failures < CONFIG.retry_count {
         if failures > 0 {
@@ -125,7 +120,7 @@ async fn main_loop() -> ExitCode {
             //TODO health check
         }
 
-        if let Err(e) = process_tasks(obf_cache.clone(),  report_cache.clone(), &mut seen_tasks).await {
+        if let Err(e) = process_tasks(&mut task_queue, &mut seen_tasks).await {
             warn!("Encountered the following error, while processing tasks: {e}");
             //failures += 1;
         } else {
@@ -139,73 +134,9 @@ async fn main_loop() -> ExitCode {
     ExitCode::from(22)
 }
 
-async fn process_task(
-    task: &BeamTask,
-    obf_cache: Shared<ObfCache>,
-    report_cache: Shared<ReportCache>,
-) -> Result<BeamResult, FocusError> {
-    debug!("Processing task {}", task.id);
-
-    let metadata: Metadata = serde_json::from_value(task.metadata.clone()).unwrap_or(Metadata {
-        project: "default_obfuscation".to_string(),
-        execute: true,
-    });
-
-    if metadata.project == "exporter" {
-        let body = &task.body;
-        return Ok(run_exporter_query(task, body, metadata.execute).await)?;
-    }
-
-    if CONFIG.endpoint_type == config::EndpointType::Blaze {
-        let query = parse_blaze_query(task)?;
-        if query.lang == "cql" {
-            // TODO: Change query.lang to an enum
-
-            Ok(run_cql_query(task, &query, obf_cache, report_cache, metadata.project).await)?
-        } else {
-            warn!("Can't run queries with language {} in Blaze", query.lang);
-            Ok(beam::beam_result::perm_failed(
-                CONFIG.beam_app_id_long.clone(),
-                vec![task.from.clone()],
-                task.id,
-                format!(
-                    "Can't run queries with language {} and/or endpoint type {}",
-                    query.lang, CONFIG.endpoint_type
-                ),
-            ))
-        }
-    } else if CONFIG.endpoint_type == config::EndpointType::Omop {
-        let decoded = decode_body(task)?;
-        let intermediate_rep_query: intermediate_rep::IntermediateRepQuery =
-            from_slice(&decoded).map_err(|e| FocusError::ParsingError(e.to_string()))?;
-        //TODO check that the language is ast
-        let query_decoded = general_purpose::STANDARD
-            .decode(intermediate_rep_query.query)
-            .map_err(FocusError::DecodeError)?;
-        let ast: ast::Ast =
-            from_slice(&query_decoded).map_err(|e| FocusError::ParsingError(e.to_string()))?;
-
-        Ok(run_intermediate_rep_query(task, ast).await)?
-    } else {
-        warn!(
-            "Can't run queries with endpoint type {}",
-            CONFIG.endpoint_type
-        );
-        Ok(beam::beam_result::perm_failed(
-            CONFIG.beam_app_id_long.clone(),
-            vec![task.from.clone()],
-            task.id,
-            format!(
-                "Can't run queries with endpoint type {}",
-                CONFIG.endpoint_type
-            ),
-        ))
-    }
-}
 
 async fn process_tasks(
-    obf_cache: Shared<ObfCache>,
-    report_cache: Shared<ReportCache>,
+    task_queue: &mut TaskQueue,
     seen: &mut HashSet<MsgId>,
 ) -> Result<(), FocusError> {
     debug!("Start processing tasks...");
@@ -216,88 +147,17 @@ async fn process_tasks(
             continue;
         }
         seen.insert(task.id);
-        let local_report_cache = report_cache.clone();
-        let local_obf_cache = obf_cache.clone();
-        tokio::spawn(handle_beam_task(task, local_obf_cache, local_report_cache));
+        task_queue.send(task).await.expect("Receiver is never dropped");
     }
     Ok(())
 }
 
-async fn handle_beam_task(task: TaskRequest<String>, local_obf_cache: Shared<ObfCache>, local_report_cache: Shared<ReportCache>) {
-    let task_cloned = task.clone();
-    let claiming = tokio::task::spawn(async move { beam::claim_task(&task_cloned).await });
-    let res = process_task(&task, local_obf_cache, local_report_cache).await;
-    let error_msg = match res {
-        Err(FocusError::DecodeError(_)) | Err(FocusError::ParsingError(_)) => {
-            Some("Cannot parse query".to_string())
-        }
-        Err(FocusError::LaplaceError(_)) => Some("Cannot obfuscate result".to_string()),
-        Err(ref e) => Some(format!("Cannot execute query: {}", e)),
-        Ok(_) => None,
-    };
-
-    let res = res.ok();
-    // Make sure that claiming the task is done before we update it again.
-    match claiming.await.unwrap() {
-        Ok(_) => {}
-        Err(FocusError::ConfigurationError(s)) => {
-            error!("FATAL: Unable to report back to Beam due to a configuration issue: {s}");
-        }
-        Err(FocusError::UnableToAnswerTask(e)) => {
-            warn!("Unable to report claimed task to Beam: {e}");
-        }
-        Err(e) => {
-            warn!("Unknown error reporting claimed task back to Beam: {e}");
-        }
-    }
-
-    const MAX_TRIES: u32 = 3600;
-    for attempt in 0..MAX_TRIES {
-        let comm_result = if let Some(ref err_msg) = error_msg {
-            beam::fail_task(&task, err_msg).await
-        } else {
-            beam::answer_task(task.id, res.as_ref().unwrap()).await
-        };
-        match comm_result {
-            Ok(_) => break,
-            Err(FocusError::ConfigurationError(s)) => {
-                error!(
-                    "FATAL: Unable to report back to Beam due to a configuration issue: {s}"
-                );
-            }
-            Err(FocusError::UnableToAnswerTask(e)) => {
-                warn!("Unable to report task result to Beam: {e}. Retrying (attempt {attempt}/{MAX_TRIES}).");
-            }
-            Err(e) => {
-                warn!("Unknown error reporting task result back to Beam: {e}. Retrying (attempt {attempt}/{MAX_TRIES}).");
-            }
-        };
-        tokio::time::sleep(Duration::from_secs(2)).await;
-    }
-}
-
-fn decode_body(task: &BeamTask) -> Result<Vec<u8>, FocusError> {
-    let decoded = general_purpose::STANDARD
-        .decode(&task.body)
-        .map_err(FocusError::DecodeError)?;
-
-    Ok(decoded)
-}
-
-fn parse_blaze_query(task: &BeamTask) -> Result<blaze::CqlQuery, FocusError> {
-    let decoded = decode_body(task)?;
-
-    let query: blaze::CqlQuery =
-        from_slice(&decoded).map_err(|e| FocusError::ParsingError(e.to_string()))?;
-
-    Ok(query)
-}
 
 async fn run_cql_query(
     task: &BeamTask,
     query: &CqlQuery,
-    obf_cache: Shared<ObfCache>,
-    report_cache: Shared<ReportCache>,
+    obf_cache: Arc<Mutex<ObfCache>>,
+    report_cache: Arc<Mutex<ReportCache>>,
     project: String,
 ) -> Result<BeamResult, FocusError> {
     let encoded_query =
@@ -443,9 +303,7 @@ fn replace_cql_library(mut query: CqlQuery) -> Result<CqlQuery, FocusError> {
             query.lib
         )))?;
 
-    let decoded_cql = general_purpose::STANDARD
-        .decode(old_data_string)
-        .map_err(FocusError::DecodeError)?;
+    let decoded_cql = util::base64_decode(old_data_string)?;
 
     let decoded_string = str::from_utf8(&decoded_cql)
         .map_err(|_| FocusError::ParsingError("CQL query was invalid".into()))?;
@@ -461,7 +319,7 @@ fn replace_cql_library(mut query: CqlQuery) -> Result<CqlQuery, FocusError> {
     };
 
     let replaced_cql_str = util::replace_cql(decoded_string);
-    let replaced_cql_str_base64 = general_purpose::STANDARD.encode(replaced_cql_str);
+    let replaced_cql_str_base64 = BASE64.encode(replaced_cql_str);
     let new_data_value = serde_json::to_value(replaced_cql_str_base64)
         .expect("unable to turn base64 string into json value - this should not happen");
 
@@ -472,7 +330,7 @@ fn replace_cql_library(mut query: CqlQuery) -> Result<CqlQuery, FocusError> {
 }
 
 fn beam_result(task: BeamTask, measure_report: String) -> Result<BeamResult, FocusError> {
-    let data = general_purpose::STANDARD.encode(measure_report.as_bytes());
+    let data = BASE64.encode(measure_report.as_bytes());
     Ok(beam::beam_result::succeeded(
         CONFIG.beam_app_id_long.clone(),
         vec![task.from],
@@ -481,6 +339,7 @@ fn beam_result(task: BeamTask, measure_report: String) -> Result<BeamResult, Foc
     ))
 }
 
+#[cfg(test)]
 mod test {
     use super::*;
